Fix: Use comint to send commands to ans-term instance
jupy-launch seemed to break when I switched to emacs 26. Sending commands
to the running ansi-term instance through (comint-send-string) restores
previous functionality. As I understand emacs shells are built with
comint tools so this might be a more robust way. In addition
ein:notebooklist-login seems to do the right thing and authenticates itself.
No need for the ultra hacky ein-get-psswd-login function that would scrape
the token from the terminal buffer.